Sonar Issue fixes
